refactor(modularity): remove duplicate domain/state_machine.py
Eliminates dead duplicate of models/state_machine.py found during
Phase 1b re-export cleanup (issue #2391).

Changes:
- Delete src/vibe3/domain/state_machine.py (41-line duplicate)
- Redirect test imports to canonical location: vibe3.models.state_machine
- All 21 affected tests pass

Rationale:
- Violates modularity: same symbols in two modules
- Tests bypassed public API by importing from submodules
- Reduces maintenance burden (no more sync burden)

Closes #2391
